+++ examples/with-emailpassword/src/Home/CallAPIView.js
@@ -7,7 +7,7 @@ export default function CallAPIView() {
     async function callAPIClicked() {
         // this will also automatically refresh the session if needed
         let response = await axios.get(getApiDomain() + "/sessioninfo");
-         console.log(getApiDomain() + "/sessioninfo")
+        console.log(getApiDomain() + "/sessioninfo");
         window.alert("Session Information:\n" + JSON.stringify(response.data, null, 2));
     }